+++ scan_json/changelog.md
@@ -2,6 +2,9 @@
 
 - Allow triggering on arrays
 - Allow triggering on basic values (strings, numbers, booleans, null)
+- Trigger on all objects, not only on unnamed in array context
+- Add `IOError` to `ScanError`
+- Add `idtransform` and `copy_atom`
 
 
 ## [1.0.2] - 2025-05-03